+++ source_gen/source_gen/source_gen/test/builder_test.dart
@@ -315,8 +315,6 @@ part "a.foo.dart";'''
           'generated/test_lib.foo.dart must be included as a part directive in the input '
               'library with:\n    part \'generated/test_lib.foo.dart\';'
         ]);
-      }, onPlatform: const {
-        'windows': Skip('https://github.com/dart-lang/source_gen/issues/570')
       });
 
       test('generates relative `path of` for output in different directory',
@@ -342,8 +340,6 @@ part "a.foo.dart";'''
               '$_pkgName|lib/generated/a.foo.dart':
                   decodedMatches(startsWith("part of '../a.dart';")),
             });
-      }, onPlatform: const {
-        'windows': Skip('https://github.com/dart-lang/source_gen/issues/570')
       });
 
       test(
@@ -672,8 +668,6 @@ foo generated content
 ''')),
           },
         );
-      }, onPlatform: const {
-        'windows': Skip('https://github.com/dart-lang/source_gen/issues/570')
       });
 
       test('warns about missing part statement', () async {
@@ -698,8 +692,6 @@ foo generated content
             'input library with:\n    part \'generated/a.g.dart\';',
           ),
         );
-      }, onPlatform: const {
-        'windows': Skip('https://github.com/dart-lang/source_gen/issues/570')
       });
     });
   });